feat: implement conditional compilation for platform backends and add Win32 window support
