+++ src/core/platform/mod.rs
@@ -1,4 +1,6 @@
 // This module contains low-level, platform-specific implementation details.
 
 pub mod direct2d_context;
+pub mod win32_window;
+pub mod window_backend;
 pub mod wndproc;
