+++ src/core/platform/win32_window.rs
@@ -1,0 +1,104 @@
+use crate::core::event::event_handler::EventHandler;
+use crate::core::platform::direct2d_context::Direct2DContext;
+use crate::core::platform::window_backend::WindowBackend;
+use crate::core::platform::wndproc::wndproc;
+use crate::core::window::config::WindowConfig;
+use windows::{
+    core::*,
+    Win32::Foundation::{GetLastError, *},
+    Win32::Graphics::Gdi::*,
+    Win32::System::LibraryLoader::GetModuleHandleW,
+    Win32::UI::WindowsAndMessaging::*,
+};
+
+/// The Win32 implementation of the `WindowBackend` trait.
+pub struct Win32Window<T, E: EventHandler<T>> {
+    pub hwnd: HWND,
+    pub d2d_context: Direct2DContext,
+    pub event_handler: E,
+    pub app: T,
+}
+
+impl<T: 'static, E: EventHandler<T> + 'static> Win32Window<T, E> {
+    /// Creates a new Win32 window.
+    pub fn new(config: &WindowConfig, event_handler: E, app: T) -> Result<Box<Self>> {
+        let instance = unsafe { GetModuleHandleW(None)? };
+        Self::register_class(instance.into(), &config.class_name)?;
+
+        let mut window = Box::new(Self {
+            hwnd: HWND(std::ptr::null_mut()),
+            d2d_context: Direct2DContext::new(&config.font_face_name, config.font_size as f32)?,
+            event_handler,
+            app,
+        });
+
+        let hwnd = unsafe {
+            CreateWindowExW(
+                WINDOW_EX_STYLE::default(),
+                &HSTRING::from(config.class_name.as_str()),
+                &HSTRING::from(config.title.as_str()),
+                WS_OVERLAPPEDWINDOW,
+                CW_USEDEFAULT,
+                CW_USEDEFAULT,
+                config.width,
+                config.height,
+                None,
+                None,
+                Some(instance.into()),
+                Some(window.as_mut() as *mut _ as *mut _),
+            )?
+        };
+
+        window.hwnd = hwnd;
+        window.d2d_context.create_device_dependent_resources(hwnd)?;
+
+        unsafe {
+            let _ = ShowWindow(hwnd, SW_SHOW);
+        };
+        unsafe {
+            let _ = UpdateWindow(hwnd);
+        };
+
+        Ok(window)
+    }
+
+    fn register_class(instance: HINSTANCE, class_name: &str) -> Result<()> {
+        let class_name_hstring = HSTRING::from(class_name);
+
+        let wc = WNDCLASSEXW {
+            cbSize: std::mem::size_of::<WNDCLASSEXW>() as u32,
+            style: CS_HREDRAW | CS_VREDRAW,
+            lpfnWndProc: Some(wndproc::<T, E>),
+            cbClsExtra: 0,
+            cbWndExtra: std::mem::size_of::<*mut Self>() as i32,
+            hInstance: instance,
+            hIcon: unsafe { LoadIconW(None, IDI_APPLICATION)? },
+            hCursor: unsafe { LoadCursorW(None, IDC_ARROW)? },
+            hbrBackground: unsafe { HBRUSH(GetStockObject(BLACK_BRUSH).0) },
+            lpszMenuName: PCWSTR::null(),
+            lpszClassName: PCWSTR::from_raw(class_name_hstring.as_ptr()),
+            hIconSm: unsafe { LoadIconW(None, IDI_APPLICATION)? },
+        };
+
+        unsafe {
+            if RegisterClassExW(&wc) == 0 {
+                return Err(Error::from_hresult(HRESULT::from_win32(GetLastError().0)));
+            }
+        }
+
+        Ok(())
+    }
+}
+
+impl<T, E: EventHandler<T>> WindowBackend<T, E> for Win32Window<T, E> {
+    fn run(&self) -> Result<()> {
+        let mut message = MSG::default();
+        while unsafe { GetMessageW(&mut message, None, 0, 0) }.into() {
+            unsafe {
+                let _ = TranslateMessage(&message);
+            };
+            unsafe { DispatchMessageW(&message) };
+        }
+        Ok(())
+    }
+}
